refactored code; added nice loading and error handlers
