+++ .gitignore
@@ -9,6 +9,9 @@ dist/
 web-build/
 expo-env.d.ts
 
+# Local
+.local/
+
 # Native
 *.orig.*
 *.jks
