+++ App.tsx
@@ -1,27 +1,17 @@
-import { Pressable, Vibration } from "react-native";
+import { useEffect } from "react";
+import { NavigationContainer } from "@react-navigation/native";
 import {
-  createBottomTabNavigator,
-  BottomTabBarButtonProps,
-} from "@react-navigation/bottom-tabs";
-import { NavigationContainer, useNavigation } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
+  createNativeStackNavigator,
+} from "@react-navigation/native-stack";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { AuthProvider } from "@/client/contexts";
+import { ProtectedRoute } from "@/client/components";
+import { MainTabNavigator } from "@/client/components/App";
 import {
-  Home,
-  Explore,
-  Add,
-  Marketplace,
-  Profile,
   SignUp,
-  ForgotPassword,
+  Login,
+  Add,
 } from "@/client/screens";
-
-import {
-  TabBarHomeIcon,
-  TabBarExploreIcon,
-  TabBarAddIcon,
-  TabBarMarketplaceIcon,
-  TabBarProfileIcon,
-} from "@/client/assets/icons/expo-icons";
 import {
   useFonts as useInterFonts,
   Inter_400Regular,
@@ -43,14 +33,7 @@ import {
 import * as SplashScreen from "expo-splash-screen";
 import * as Sentry from "@sentry/react-native";
 import "./global.css";
-import { useEffect } from "react";
-import { RootStackParamList } from "@/types/AppModels";
-import { NativeStackNavigationProp } from "@react-navigation/native-stack";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import "react-native-url-polyfill/auto";
-import { AuthProvider } from "@/client/contexts/AuthContext";
-import Login from "@/client/screens/Login";
-import { ProtectedRoute } from "@/client/components/ProtectedRoute";
 
 if (!process.env.EXPO_PUBLIC_SENTRY_DSN) {
   throw new Error("SENTRY_DSN is not set");
@@ -70,112 +53,7 @@ Sentry.mobileReplayIntegration({
   maskAllVectors: true,
 });
 
-const Tab = createBottomTabNavigator();
 const Stack = createNativeStackNavigator();
-
-const TabBarButton = (props: BottomTabBarButtonProps) => (
-  <Pressable
-    {...props}
-    onPress={props.onPress}
-    android_ripple={{ color: "transparent" }}
-    style={props.style}
-  />
-);
-const AddButton = (props: BottomTabBarButtonProps) => {
-  const navigation =
-    useNavigation<NativeStackNavigationProp<RootStackParamList>>();
-  const handlePress = () => {
-    Vibration.vibrate(50);
-    navigation.navigate("Add");
-  };
-
-  return <Pressable {...props} onPress={handlePress} />;
-};
-
-const MainTabNavigator = () => {
-  return (
-    <Tab.Navigator
-      screenOptions={{
-        tabBarShowLabel: false,
-        headerShown: false,
-        tabBarStyle: {
-          height: 70,
-          paddingHorizontal: 15,
-        },
-        tabBarButton: TabBarButton,
-      }}
-      initialRouteName="Home"
-    >
-      <Tab.Screen
-        name="Home"
-        component={Home}
-        options={{
-          tabBarIcon: ({ focused }: { focused: boolean }) =>
-            TabBarHomeIcon({ focused }),
-          tabBarIconStyle: {
-            marginTop: 12,
-          },
-          animation: "shift",
-          tabBarButton: TabBarButton,
-        }}
-      />
-      <Tab.Screen
-        name="Explore"
-        component={Explore}
-        options={{
-          tabBarIcon: ({ focused }: { focused: boolean }) =>
-            TabBarExploreIcon({ focused }),
-          tabBarIconStyle: {
-            marginTop: 12,
-          },
-          tabBarButton: TabBarButton,
-        }}
-      />
-      <Tab.Screen
-        name="AddButton"
-        options={{
-          tabBarIcon: ({ focused }: { focused: boolean }) =>
-            TabBarAddIcon({ focused }),
-          tabBarIconStyle: {
-            backgroundColor: "#f0f0f0",
-            marginTop: 6,
-            width: 55,
-            height: 45,
-            borderRadius: 15,
-          },
-          tabBarButton: AddButton,
-        }}
-      >
-        {() => null}
-      </Tab.Screen>
-      <Tab.Screen
-        name="Marketplace"
-        component={Marketplace}
-        options={{
-          tabBarIcon: ({ focused }: { focused: boolean }) =>
-            TabBarMarketplaceIcon({ focused }),
-          tabBarIconStyle: {
-            marginTop: 12,
-          },
-          tabBarButton: TabBarButton,
-        }}
-      />
-      <Tab.Screen
-        name="Profile"
-        component={Profile}
-        options={{
-          tabBarIcon: ({ focused }: { focused: boolean }) =>
-            TabBarProfileIcon({ focused }),
-          tabBarIconStyle: {
-            marginTop: 12,
-          },
-          animation: "shift",
-          tabBarButton: TabBarButton,
-        }}
-      />
-    </Tab.Navigator>
-  );
-};
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -248,16 +126,6 @@ function App() {
                 headerShown: false,
               }}
             />
-            <Stack.Screen
-              name="ForgotPassword"
-              component={ForgotPassword}
-              options={{
-                animation: "ios_from_right",
-                gestureEnabled: true,
-                gestureDirection: "horizontal",
-                headerShown: false,
-              }}
-            />
             <Stack.Screen name="Tabs" options={{ headerShown: false }}>
               {() => (
                 <ProtectedRoute>
@@ -273,9 +141,7 @@ function App() {
               }}
             >
               {() => (
-                <ProtectedRoute>
-                  <Add />
-                </ProtectedRoute>
+                <Add />
               )}
             </Stack.Screen>
           </Stack.Navigator>
