+++ app/client/components/App/TabBarComponents.tsx
@@ -1,0 +1,4 @@
+import TabBarButton from "./TabBarButton";
+import AddButton from "./AddButton";
+
+export { TabBarButton, AddButton };
