+++ app/client/components/App/index.ts
@@ -1,0 +1,3 @@
+export { default as MainTabNavigator } from "./MainTabNavigator";
+export { default as TabBarButton } from "./TabBarButton";
+export { default as AddButton } from "./AddButton";
