+++ app/client/components/ProtectedRoute.tsx
@@ -8,7 +8,7 @@ type ProtectedRouteProps = {
   children: React.ReactNode;
 };
 
-export function ProtectedRoute({ children }: ProtectedRouteProps) {
+function ProtectedRoute({ children }: ProtectedRouteProps) {
   const { session, isLoading } = useAuth();
   const navigation =
     useNavigation<NativeStackNavigationProp<RootStackParamList>>();
@@ -32,3 +32,5 @@ export function ProtectedRoute({ children }: ProtectedRouteProps) {
 
   return <>{children}</>;
 }
+
+export default ProtectedRoute;
