+++ app/client/components/index.ts
@@ -5,3 +5,8 @@ export { default as Title } from "./Title";
 export { default as Checkbox } from "./Checkbox";
 export { default as Radio } from "./Radio";
 export { default as Label } from "./Label";
+export { default as LoginForm } from "./Login/LoginForm";
+export { default as SetupFormWithZod } from "./SignUp/SetupFormStep";
+export { default as QRCodeStep } from "./SignUp/QRStep";
+export { default as NotificationsStep } from "./SignUp/NotificationsStep";
+export { default as ProtectedRoute } from "./ProtectedRoute";
