+++ app/client/contexts/index.ts
@@ -1,0 +1,1 @@
+export { default as AuthProvider } from "./AuthContext";
