+++ app/client/hocs/index.ts
@@ -1,0 +1,1 @@
+export { default as WithZodForm } from "./WithZodForm";
