+++ app/client/screens/Loading.tsx
@@ -1,0 +1,11 @@
+import { View, Text } from "react-native";
+
+const Loading = () => {
+  return (
+    <View>
+      <Text>Loading...</Text>
+    </View>
+  );
+};
+
+export default Loading;
