+++ app/client/screens/Login.tsx
@@ -1,112 +1,100 @@
-import { View, Text, TouchableOpacity, ScrollView, Image } from "react-native";
-import Title from "@/client/components/Title";
-import InputField from "@/client/components/Input";
-import { withZodForm, InjectedProps } from "@/client/hocs/WithZodForm";
-import { LoginFormType } from "@/types/FormModels";
-import { LOGIN_FORM_MODULE } from "@/utils/constants";
-import Button from "@/client/components/Button";
-// import GoogleIconSvg from "@/client/assets/icons/google-icon.svg";
+import { useState, useEffect } from "react";
+import {
+  ScrollView,
+  Image,
+  Text,
+  TouchableOpacity,
+  View,
+  Dimensions,
+} from "react-native";
 import { useNavigation } from "@react-navigation/native";
-import { RootStackParamList } from "@/types/AppModels";
+import { AppErrorCodes, RootStackParamList } from "@/types/AppModels";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
-import { useSignIn } from "@/client/hooks/useAuth";
-
-const LoginFormComponent = ({
-  form,
-  handleFormSubmit,
-}: InjectedProps<LoginFormType>) => {
-  const { control } = form;
-  // const navigation = useNavigation<NativeStackNavigationProp<RootStackParamList>>();
-
-  return (
-    <View>
-      <View>
-        <InputField
-          label="Email address or username"
-          id="emailOrUsername"
-          name="emailOrUsername"
-          control={control}
-          autoComplete="email"
-          inputMode="email"
-        />
-      </View>
-      <View>
-        <InputField
-          label="Password"
-          id="password"
-          name="password"
-          control={control}
-          type="password"
-          autoComplete="password"
-        />
-      </View>
-      {/* <View className="mt-4">
-          <TouchableOpacity
-            onPress={() => navigation.navigate("ForgotPassword")}
-            className="w-full flex justify-end"
-          >
-            <Title variant="extra-small" className="text-primary-600">
-              Forgot password?
-            </Title>
-          </TouchableOpacity>
-        </View> */}
-      <View className="mt-4">
-        <Button
-          text="Sign in"
-          variant="primary"
-          onPress={handleFormSubmit}
-          className="mx-auto"
-        />
-      </View>
-    </View>
-  );
-};
-
-const LoginForm = withZodForm<LoginFormType>(LoginFormComponent);
+import { auth } from "@/client/hooks";
+import { SceneMap, TabView } from "react-native-tab-view";
+import { LoginForm, Title } from "@/client/components";
+import { LOGIN_FORM_MODULE } from "@/utils/constants";
+import LoadingScreen from "./Loading";
+import { handleError } from "@/utils/helpers";
+import { LoginFormType } from "@/types/FormModels";
 
 export default function LoginScreen() {
+  const [currentStep, setCurrentStep] = useState(0);
   const navigation =
     useNavigation<NativeStackNavigationProp<RootStackParamList>>();
-  const { mutate: signIn } = useSignIn();
+  const {
+    mutate: signIn,
+    isPending: isSigningIn,
+    isError: isSignInError,
+  } = auth.useSignIn();
+
+  useEffect(() => {
+    if (isSigningIn) {
+      setCurrentStep(1);
+    }
+    if (isSignInError) {
+      setCurrentStep(0);
+    }
+  }, [isSigningIn, isSignInError]);
 
   return (
-    <ScrollView className="flex min-h-full flex-1 flex-col bg-background relative isolate">
-      <View className="mt-28">
-        <Image
-          source={require("@/client/assets/logo.png")}
-          className="w-32 h-32 mx-auto"
-        />
-        <Title
-          className="text-center w-full text-foreground"
-          variant="extra-large"
-        >
-          Sign in to your account
-        </Title>
-      </View>
+    <TabView
+      onIndexChange={setCurrentStep}
+      initialLayout={{
+        width: Dimensions.get("window").width,
+        height: Dimensions.get("window").height,
+      }}
+      swipeEnabled={false}
+      renderTabBar={() => null}
+      navigationState={{
+        index: currentStep,
+        routes: [
+          { key: "login", title: "Login" },
+          { key: "loading", title: "Loading" },
+        ],
+      }}
+      renderScene={SceneMap({
+        login: () => (
+          <ScrollView className="flex min-h-full flex-1 flex-col bg-background relative isolate">
+            <View className="mt-28">
+              <Image
+                source={require("@/client/assets/logo.png")}
+                className="w-32 h-32 mx-auto"
+              />
+              <Title
+                className="text-center w-full text-foreground"
+                variant="extra-large"
+              >
+                Sign in to your account
+              </Title>
+            </View>
 
-      <View className="mx-auto w-full">
-        <View className="bg-white px-6 py-12">
-          <View className="space-y-6">
-            <LoginForm
-              defaultValues={LOGIN_FORM_MODULE.defaultValues}
-              validationSchema={LOGIN_FORM_MODULE.validationSchema}
-              onSubmit={(data) => {
-                signIn(data, {
-                  onSuccess: () => {
-                    navigation.reset({
-                      index: 0,
-                      routes: [{ name: "Tabs" }],
-                    });
-                  },
-                  onError: (error) => {
-                    console.error("Login failed:", error);
-                  },
-                });
-              }}
-            />
-          </View>
-        </View>
-        {/* <View>
+            <View className="mx-auto w-full">
+              <View className="bg-white px-6 py-12">
+                <View className="space-y-6">
+                  <LoginForm
+                    defaultValues={LOGIN_FORM_MODULE.defaultValues}
+                    validationSchema={LOGIN_FORM_MODULE.validationSchema}
+                    onSubmit={(data: LoginFormType) => {
+                      signIn(data, {
+                        onSuccess: () => {
+                          navigation.reset({
+                            index: 0,
+                            routes: [
+                              { name: "Tabs", params: { screen: "Profile" } },
+                            ],
+                          });
+                        },
+                        onError: (error: unknown) => {
+                          console.error("Login failed:", error);
+                          handleError(error as AppErrorCodes);
+                        },
+                      });
+                    }}
+                  />
+                </View>
+              </View>
+              {/* <View>
               <View className="relative mt-10 h-4 flex items-center justify-center">
                 <View className="absolute top-1/2 w-full h-px bg-secondary-200" />
                 <View className="z-10 px-2">
@@ -115,8 +103,8 @@ export default function LoginScreen() {
                   </Text>
                 </View>
               </View> */}
-        {/* // TODO: Add Google sign in */}
-        {/* <View className="mt-6 flex flex-row ">
+              {/* // TODO: Add Google sign in */}
+              {/* <View className="mt-6 flex flex-row ">
                 <Button
                   text="Google"
                   variant="secondary"
@@ -128,20 +116,23 @@ export default function LoginScreen() {
                   }}
                 />
               </View> */}
-        {/* </View>
+              {/* </View>
           </View> */}
-
-        <View className="flex-row justify-center items-center">
-          <Text className="text-center text-md font-inter text-secondary-500">
-            Not a member?{" "}
-          </Text>
-          <TouchableOpacity onPress={() => navigation.navigate("SignUp")}>
-            <Text className="font-semibold text-md text-primary-600 font-geist">
-              Sign up
-            </Text>
-          </TouchableOpacity>
-        </View>
-      </View>
-    </ScrollView>
+              <View className="flex-row justify-center items-center">
+                <Text className="text-center text-md font-inter text-secondary-500">
+                  Not a member?{" "}
+                </Text>
+                <TouchableOpacity onPress={() => navigation.navigate("SignUp")}>
+                  <Text className="font-semibold text-md text-primary-600 font-geist">
+                    Sign up
+                  </Text>
+                </TouchableOpacity>
+              </View>
+            </View>
+          </ScrollView>
+        ),
+        loading: () => <LoadingScreen />,
+      })}
+    />
   );
 }
