+++ app/client/screens/index.ts
@@ -4,4 +4,4 @@ export { default as Add } from "./Add";
 export { default as Marketplace } from "./Marketplace";
 export { default as Profile } from "./Profile";
 export { default as SignUp } from "./SignUp";
-export { default as ForgotPassword } from "./ForgotPassword";
+export { default as Login } from "./Login";
