+++ tsconfig.json
@@ -21,5 +21,10 @@
     }
   },
   "exclude": ["node_modules", "babel.config.js", "metro.config.js"],
-  "include": ["**/*.ts", "**/*.tsx", "nativewind-env.d.ts"]
+  "include": [
+    "**/*.ts",
+    "**/*.tsx",
+    "nativewind-env.d.ts",
+    ".local/ForgotPassword.tsx"
+  ]
 }
